+++ src/main/java/com/example/demo/service/UserService.java
@@ -5,7 +5,7 @@ import com.example.demo.entity.User;
 public interface UserService {
     public User getUser(long id);
 
-    public int add(User user);
+    public Long add(User user);
 
     int update(User user);
 }
